Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.

Predicts the probability that a x3p scan will be good and also predicts the reason why it is bad quality if it is predicted to be bad quality.

predict_quality(x3pobjectvector, x3pnamevector)

Arguments

x3pobjectvector

a vector of x3p objects to be classified as good or bad quality scans

x3pnamevector

a vector of names of x3p files

Value

a dataframe containing a vector of predicted probabilities of a scan being good and its predicted quality_type

Examples

data(fau277_bb_l2)
predict_quality(list(fau277_bb_l2, fau277_bb_l2), x3ptools::x3p_show_xml(fau277_bb_l2,"Comment")$Comment)
#> Error in stopifnot(class(x3p) == "x3p"): object 'x3pobjectliist' not found